requirements.txt
kopf>=1.36.0
kubernetes>=28.1.0
httpx>=0.25.0
pydantic>=2.5.0
pyyaml>=6.0.0
asyncio-throttle>=1.0.2
